web/requirements.txt
wheel
Flask
flask_restful
flask_jwt_extended==3.24.1
flask_cors
flask_sockets
sentry-sdk[flask]
raven
gevent
gevent-websocket
greenlet
pyJWT
hyper
pymongo
bcrypt
gunicorn
Jinja2
Werkzeug==1.0.1
boto3==1.18.60
pytest
pytest-gevent
pytest-ordering
rq
rq_dashboard
rq-retry
redis
newrelic
pillow
requests
elasticsearch==7.12.1
requests-aws4auth
twilio==6.27.1
python-dotenv
phonenumbers
dnspython
mailchimp-marketing
instaloader
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
six>=1.13.0
apscheduler
requests-oauthlib
stripe==2.56.0
webargs==8.0.0
slack_sdk
playwright==1.14.1
TikTokApi
selenium
webdriver-manager
nordvpn-connect==0.0.7
nordvpn-switcher==0.2.7
